refactor: address issues raised in PR
